requirements.txt
bcrypt==4.0.1
coverage==7.2.7
cryptography==3.4.8
Django==4.2.1
django-cors-headers==4.1.0
django-debug-toolbar==4.1.0
django-filter==23.2
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
gunicorn==20.1.0
model-bakery==1.12.0
numpy
psycopg2-binary==2.9.6
pytest==7.3.1
pytest-cov==4.1.0
pytest-django==4.5.2
python-decouple==3.8
PyYAML
requests==2.31.0
sqlparse==0.4.4
celery
redis
drf-spectacular
social-auth-app-django
social-auth-app-django-mongoengine
django-silk